Expect virtual addresses for the buffers in UHYVE_PORT_READ/UHYVE_PORT_WRITE and perform the address translation in uhyve
Depending on the size of the data to be read/written, the buffers could cross page boundaries in HermitCore.
Previously, uhyve expected the physical addresses of the pages behind the buffers to be contiguous.
However, the OS only guarantees that the virtual addresses for the buffer memory are contiguous.
On the contrary, each virtual address may map to a physical address anywhere in memory.